Lab 2 Part 2 Try 2
